Refactor CLI implementation to load table in main
A clap CLI with subcommands can't have a global required argument.
The argument must either be optional (which the URI isn't) or be
passed as argument to each subcommand. Previous implementation matched
each sub-command's explicit arguments. The new implementation abstracts
the each sub-command's arguments to its own struct.

The `Command` enum also implement a `get_uri` to read the (required)
URI argument passed to any of the sub-commands.