Add feature flag 'optimize' to avoid datafusion
The datafusion crate adds a lot of additional code to the final binary:
* debug: 348M -> 802M
* release: 24M -> 64M

This is only necessary when one wants to run compaction or z-ordering
on Deltalake tables. Some use cases may only be interested in the
read features of `deltactl`. The feature flag allows those use cases
to avoid the added size and build time.