Move table URI cli arg to global option
It's not possible to have global required options in `clap`.
The alternatives are:

1. Identical required argument on each sub-command (as currently);
2. A global non-required option (as in this commit).

The URI is really required to do anything meaningful in the program,
but the implementation to achieve alt. 1 is overy complex. A simpler
solution is alt. 2, with an explicit early exit when the URI is not
provided.